// src/lib/api/schemas/RejectGuideRequest.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * Elite API
 * A backend API for https://elitebot.dev/ that provides Hypixel Skyblock data.
<br><br>
Use of this API requires following the [Elite API TOS](https://elitebot.dev/apiterms). This API is not affiliated with Hypixel or Mojang.
 * OpenAPI spec version: v1
 */

export interface RejectGuideRequest {
	/**
	 * Optional reason for rejection to provide feedback to the author.
	 * @nullable
	 */
	reason?: string | null;
}