// src/lib/api/schemas/UpdateTagRequest.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * Elite API
 * A backend API for https://elitebot.dev/ that provides Hypixel Skyblock data.
<br><br>
Use of this API requires following the [Elite API TOS](https://elitebot.dev/apiterms). This API is not affiliated with Hypixel or Mojang.
 * OpenAPI spec version: v1
 */

export interface UpdateTagRequest {
	/** @nullable */
	name?: string | null;
	/** @nullable */
	category?: string | null;
	/** @nullable */
	hexColor?: string | null;
}